requirements.txt
Flask
flask_sqlalchemy
flask_migrate
flask-admin
flask-login
flask-dance
phonenumbers
authlib
psycopg2
psycopg2-binary
python-dotenv
requests
httpx
gotrue